Refactored code for creating DataLakeFileSystemClient mocks into a base class for reuse
